requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.1
docopt==0.6.2
gitignore==0.0.8
python-decouple==3.5
sqlparse==0.4.2
tzdata==2021.5
